Minor shuffling of code and add some documentation
