+++ kv/tikv/inner_server/modify.go
@@ -1,5 +1,6 @@
 package inner_server
 
+// ModifyType is the smallest unit of mutation of TinyKV's underlying storage (i.e., raw key/values on disk(s))
 type ModifyType int64
 
 const (
